修复uD分析测试脚本并添加副本测试
- test_ud_analysis_copy.py: 修正results赋值逻辑(字典赋值替代覆盖)
  和tag大小写一致性问题
- test_ud_analysis.py: 同步修复print缩进和变量名问题
- super_config.h/c: 重新生成

+++ frameworkCodes/c/super_config.h
@@ -44,7 +44,7 @@ typedef struct {
     REAL cmd_speed;
     REAL cmd_id;
     REAL cmd_iq;
-    long cmd_ud;
+    REAL cmd_ud;
     REAL cmd_uq;
     REAL tload;
     REAL tload_time;
